refactor(llm): extract KvSlotManager -- multi-slot KV state machine (stage 3b)
Move the ~450-line multi-slot prefix-KV state machine out of LLM::Impl into
src/runner/KvSlotManager.hpp: init_kv_slots, host_dump/host_load_kv, save/activate,
select_kv_slot(_by_history), commit_kv_slot_choice, plus the 6 slot-state members.
Impl now holds a KvSlotManager and implements a narrow IKvSlotHost interface
(distinct slot_* forwarders so Impl's hot-path helpers stay non-virtual; the 7
per-conversation decode-state members are snapshotted via slot_capture/restore).
Backend mem-copy helpers + LLM_* addr macros moved to LLMLayer.hpp so the manager
can touch device KV. Only 2 external reads in Run() (multi_slot_enabled/active)
became accessor calls; hash_active_kv (the KV fingerprint) stays in Impl.

LLM.cpp: 5720 -> 5140 lines (this stage -269; whole effort -580).

Verified: LLM.cpp compiles on BOTH build_x86 (AXCL) and build_650 (on-chip);
golden_runner built WITH this change --compare vs the 5 pre-3b baselines on
AX650N -> ALL PASS (byte-identical output + KV-content fingerprint across
qwen3-0.6b/1.7b, smollm2-360m, gemma3 multi-turn + qwen3-0.6b interleaved
multi-slot select/evict). Behavior-preserving.
